requirements.txt
sqlalchemy==2.0.19
psycopg2-binary==2.9.7
pandas==2.3.0
python-dotenv==1.1.0
together==1.4.6
black==25.1.0